+++ src/icemac/ab/calexport/browser/tests/test_renderer.py
@@ -90,3 +90,18 @@ def test_renderer__ExportEvent__dd_class__3(
     """
     revent = SpecialFieldRecurredEventFactory(address_book, with_field=False)
     assert None is ExportEventFactory(revent).dd_class()
+
+
+def test_renderer__ExportEvent__action_url__1(
+        address_book, URLFieldRecurredEventFactory, ExportEventFactory):
+    """It returns `None` by default."""
+    revent = URLFieldRecurredEventFactory(address_book)
+    assert None is ExportEventFactory(revent).action_url()
+
+
+def test_renderer__ExportEvent__action_url__2(
+        address_book, URLFieldRecurredEventFactory, ExportEventFactory):
+    """It returns `None` by default."""
+    revent = URLFieldRecurredEventFactory(
+        address_book, with_field=True, value='http://event.info')
+    assert 'http://event.info' is ExportEventFactory(revent).action_url()